Split audio engine projection
